Validate group and expected match counts in test cases
If these are not equal, it most likely means that multiple matches are
expected int he input string, which is not what we want. We only want
the first search to find the regex pattern, as that is the only one
necessary. Then, if someone wants to find multiple instances of the
pattern in text, they can do it with e.g. re.findall using the pattern
found.